v5.x-scala-js: JitCost, ErgoAlgos, SigmaConstants moved to interpreter module
