+++ interpreter/shared/src/main/scala/org/ergoplatform/SigmaConstants.scala
@@ -1,6 +1,5 @@
 package org.ergoplatform
 
-import scalan.util.CollectionUtil._
 import sigmastate.basics.CryptoConstants
 
 case class SizeConstant[T: Numeric](value: T, id: Short, description: String)
