+++ interpreter/shared/src/main/scala/org/ergoplatform/settings/ErgoAlgos.scala
